src/requirements.txt
requests>=2.25.1,<3.0
beautifulsoup4>=4.11.1,<5.0
PyYaml>=6.0,<7.0
